DFS algorithm implemented Comments added
