+++ src/java/com/cesar/graph/Node.java
@@ -11,6 +11,12 @@ public class Node {
         this.nodeid = nodeid;
     }
 
+    /**
+     * Connects the node called to the node provided.
+     * @param node Node to connect to.
+     * @param weight Weight of the relation.
+     * @param directional Wether or not the target node should connect back.
+     */
     public void connectNode(Node node, int weight, boolean directional){
         vertices.put(node.nodeid, weight);
         if (!directional){
